01_Langchain_Models/requirements.txt
#langchain core
langchain
langchain-core

#open-ai integration
langchain-openai
openai

#Anthropic integration
langchain-anthropic

#Google Gemini (Palm) integration
langchain-google-genai
google-generativeai

#Hugging Face integration
langchain-huggingface
transformers
huggingface-hub

#environment variable management
python-dotenv

#Machine learning libraries
numpy
scikit-learn
